# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/dorm": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dorm"
                ],
                "summary": "get list of dorms",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "capacity",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "lat",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "long",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "lower_price",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "upper_price",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "zone",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "dorm facility",
                        "name": "dorm_facilities",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "room facility",
                        "name": "room_facilities",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.DormDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dorm"
                ],
                "summary": "create dorm",
                "parameters": [
                    {
                        "description": "register dorm",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterDormDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Dorm"
                        }
                    }
                }
            }
        },
...
        "/dorm/suggest/{letter}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dorm"
                ],
                "summary": "get dorm names from first letter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "First Letter",
                        "name": "letter",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.DormSuggestionDTO"
                            }
                        }
                    }
                }
            }
        },
...
        "/dorm/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dorm"
                ],
                "summary": "get dorm details from dorm id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dorm ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Dorm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dorm"
                ],
                "summary": "update dorm",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dorm ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "register dorm",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateDormDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Dorm"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dorm"
                ],
                "summary": "delete dorm",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dorm ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/dorm/{id}/picture": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dorm"
                ],
                "summary": "update dorm pictures",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dorm ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "upload multiple room pictures,test this out in postman",
                        "name": "pictures",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RoommateRequestWithUnregisteredDorm"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.DormDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "long": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "starting_price": {
                    "type": "integer"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
        "dto.DormSuggestionDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.RegisterDormDTO": {
            "type": "object",
            "required": [
                "address",
                "facilities",
                "lat",
                "long",
                "name",
                "owner_id",
                "type",
                "zone"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "facilities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "lat": {
                    "type": "number"
                },
                "long": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "rules": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateDormDTO": {
            "type": "object",
            "required": [
                "address",
                "facilities",
                "lat",
                "long",
                "name",
                "type",
                "zone"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "facilities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "lat": {
                    "type": "number"
                },
                "long": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "rules": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
...
        "model.AllDormFacility": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "model.AllRoomFacility": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.Dorm": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "facilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AllDormFacility"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "long": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "nearby_locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.NearbyLocation"
                    }
                },
                "pictures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DormPicture"
                    }
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Room"
                    }
                },
                "rules": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
...
        "model.DormPicture": {
            "type": "object",
            "properties": {
                "picture_url": {
                    "type": "string"
                }
            }
        },
...
        "model.Location": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "long": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.NearbyLocation": {
            "type": "object",
            "properties": {
                "distance": {
                    "type": "number"
                },
                "location": {
                    "$ref": "#/definitions/model.Location"
                }
            }
        },
...
        "model.Room": {
            "type": "object",
            "properties": {
                "available_from": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "dorm_id": {
                    "type": "integer"
                },
                "facilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AllRoomFacility"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pictures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RoomPicture"
                    }
                },
                "price": {
                    "type": "integer"
                },
                "size": {
                    "type": "number"
                }
            }
        },
...
        "model.RoomPicture": {
            "type": "object",
            "properties": {
                "picture_url": {
                    "type": "string"
                }
            }
        },
...
        "model.RoommateRequestUnregisteredDormPicture": {
            "type": "object",
            "properties": {
                "picture_url": {
                    "type": "string"
                }
            }
        },
...
        "model.RoommateRequestWithUnregisteredDorm": {
            "type": "object",
            "properties": {
                "dorm_name": {
                    "type": "string"
                },
                "facilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AllRoomFacility"
                    }
                },
                "number_of_roommates": {
                    "type": "integer"
                },
                "room_description": {
                    "type": "string"
                },
                "room_pictures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RoommateRequestUnregisteredDormPicture"
                    }
                },
                "room_price": {
                    "type": "integer"
                },
                "shared_room_price": {
                    "type": "integer"
                },
                "size": {
                    "type": "number"
                },
                "student_id": {
                    "type": "string"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}